% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sample.R
\name{boot_sample}
\alias{boot_sample}
\title{Function to create bootstrap samples of LDA projected means}
\usage{
boot_sample(values, N.b = 100)
}
\arguments{
\item{values}{A matrix of group ID's and the first 2 lda projections of the data.}

\item{N.b}{The number of bootstrap sample means to be estimated.}
}
\value{
returns matrix of lda projected means dim n.b*number of groups by 3, columns group, LDA1, and LDA2.
}
\description{
Bootstrap resampling used for mean CI estimation
}
\examples{
library(MASS)
# simulated data set to give random groups
q=100
data <- as.data.frame(list(x1 = runif(q), x2 = rnorm(q), x3 = rlnorm(q), group = sample(c('s','d','w'),q,replace=TRUE)))

# create lda projections, though not needed function will work with any bivariate data and corresponding grouping column
lda <- lda(group ~., data = data)

# format data, groups need to be numeric, can not be factors of characters
V1<-as.numeric(as.factor(data$group))
lda.vec<-as.data.frame(lda$scaling)
lda.p <- predict(lda)
v <- as.data.frame(cbind(V1, lda.p$x))
str(v)

# create bootstrap sample means
b <- boot_sample(values = v)
str(b)
}
\author{
James Colee
}
