% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_posterior_plot.R
\name{lda_posterior_plot}
\alias{lda_posterior_plot}
\title{Function to make LDA posterior probabilities by group}
\usage{
lda_posterior_plot(
  lda,
  groups,
  scale = FALSE,
  groups.names = NULL,
  labels = NULL,
  x.axis = NULL,
  y.axis = NULL
)
}
\value{
returns multiple plots of LDA posterior probabilities by groups
}
\description{
discribtion
}
\examples{
#posterior probability contour plot by goup
#examples
library(MASS)
q=100
data <- as.data.frame(list(x1 = runif(q), x2 = rnorm(q), x3 = rlnorm(q), group = sample(c('s','d','w'),q,replace=TRUE)))
lda <- lda(group ~., data = data) 
lda_posterior_plot(lda=lda, groups = data$group, scale = TRUE)
}
\author{
James Colee
}
