% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sample.R
\name{boot_sample}
\alias{boot_sample}
\title{Function to create bootstrap samples of LDA projected means}
\usage{
boot_sample(values, N.b = 100)
}
\value{
returns matrix of lda projected means dim n.b*number of groups by 3, columns group, LDA1, and LDA2
}
\description{
discribtion
}
\examples{
#example
library(MASS)
q <-100
data <- as.data.frame(list(x1 = runif(q), x2 = rnorm(q), x3 = rlnorm(q), group = sample(c('s','d','w'),q,replace=TRUE)))
lda <- lda(group ~., data = data) 
as.numeric(as.factor(data$group))
lda.p <- predict(lda)
v <- as.data.frame(cbind(V1, lda.p$x))
b <- boot_sample(values = v)
str(b)
}
\author{
James Colee
}
