% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_gplot.R
\name{lda_gplot}
\alias{lda_gplot}
\title{Function to make LDA plots wich confidence bounds}
\usage{
lda_gplot(
  lda,
  groups,
  alpha = 0.05,
  N.boot = 101,
  N.perm = 101,
  fill = TRUE,
  means = TRUE,
  biplot = TRUE,
  normal.CI = TRUE,
  circles = FALSE,
  boot.CI = FALSE,
  perm.CI = FALSE,
  scale = FALSE,
  groups.names = NULL,
  labels = NULL,
  x.axis = NULL,
  y.axis = NULL,
  size = 0.1
)
}
\value{
returns a ggolot2 object of the plot LDA projects with CI's
}
\description{
discribtion
}
\examples{
library(MASS)
q=100
data <- as.data.frame(list(x1 = runif(q), x2 = rnorm(q), x3 = rlnorm(q), group = sample(c('s','d','w'),q,replace=TRUE)))
lda <- lda(group ~., data = data) 
p<-lda_gplot(lda=lda, groups = data$group,perm.CI = TRUE,boot.CI=TRUE, circles = TRUE)
p
#no legend
p + theme(legend.position='none')
no legend title
p + theme(legend.title = element_blank()) +  scale_color_hue(labels = labels)
# add a theme
p+ theme_bw()
}
\author{
James Colee
}
