% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_sample.R
\name{perm_sample}
\alias{perm_sample}
\title{Function to create permutation estimated samples of LDA projected means}
\usage{
perm_sample(values, N.p = 100)
}
\value{
returns matrix of lda projected means dim n.b*number of groups by 3, columns group, LDA1, and LDA2
}
\description{
discribtion
}
\examples{
#example
library(MASS)
q=100
data <- as.data.frame(list(x1 = runif(q), x2 = rnorm(q), x3 = rlnorm(q), group = sample(c('s','d','w'),q,replace=TRUE)))
lda <- lda(group ~., data = data) 
V1<-as.numeric(as.factor(data$group))
lda.vec<-as.data.frame(lda$scaling)
lda.p <- predict(lda)
v <- as.data.frame(cbind(V1, lda.p$x))
b<-perm_sample(values = v)
str(b)
}
\author{
James Colee
}
